pyproject.toml
[project]
name = "exam-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "google-generativeai>=0.8.5",
    "pymupdf>=1.26.4",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
]
